+++ charts_flutter/example/lib/behaviors/selection_line_highlight_custom_shape.dart
@@ -1,37 +1,73 @@
-# Selection Line Highlight Behaviors Example
-
-![](selection_line_highlight_full.png)
-
-Example:
-
-```
+// Copyright 2018 the Charts project authors. Please see the AUTHORS file
+// for details.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// EXCLUDE_FROM_GALLERY_DOCS_START
 import 'dart:math';
 // EXCLUDE_FROM_GALLERY_DOCS_END
 import 'package:charts_flutter/flutter.dart' as charts;
 import 'package:flutter/material.dart';
 
-class SelectionLineHighlight extends StatelessWidget {
+class SelectionLineHighlightCustomShape extends StatelessWidget {
   final List<charts.Series> seriesList;
   final bool animate;
 
-  SelectionLineHighlight(this.seriesList, {this.animate});
+  SelectionLineHighlightCustomShape(this.seriesList, {this.animate});
 
   /// Creates a [LineChart] with sample data and no transition.
-  factory SelectionLineHighlight.withSampleData() {
-    return new SelectionLineHighlight(
+  factory SelectionLineHighlightCustomShape.withSampleData() {
+    return new SelectionLineHighlightCustomShape(
       _createSampleData(),
       // Disable animations for image tests.
       animate: false,
     );
   }
 
+  // EXCLUDE_FROM_GALLERY_DOCS_START
+  // This section is excluded from being copied to the gallery.
+  // It is used for creating random series data to demonstrate animation in
+  // the example app only.
+  factory SelectionLineHighlightCustomShape.withRandomData() {
+    return new SelectionLineHighlightCustomShape(_createRandomData());
+  }
+
+  /// Create random data.
+  static List<charts.Series<LinearSales, num>> _createRandomData() {
+    final random = new Random();
+
+    final data = [
+      new LinearSales(0, random.nextInt(100)),
+      new LinearSales(1, random.nextInt(100)),
+      new LinearSales(2, random.nextInt(100)),
+      new LinearSales(3, random.nextInt(100)),
+    ];
+
+    return [
+      new charts.Series<LinearSales, int>(
+        id: 'Sales',
+        domainFn: (LinearSales sales, _) => sales.year,
+        measureFn: (LinearSales sales, _) => sales.sales,
+        data: data,
+      )
+    ];
+  }
+  // EXCLUDE_FROM_GALLERY_DOCS_END
 
   @override
   Widget build(BuildContext context) {
-    // This is just a simple line chart with a behavior that highlights the
-    // selected points along the lines. A point will be drawn at the selected
-    // datum's x,y coordinate, and a vertical follow line will be drawn through
-    // it.
+    // This is a simple line chart with a behavior that highlights hovered
+    // lines. A hollow rectangular shape will be drawn at the hovered datum's
+    // x,y coordinate, and a vertical follow line will be drawn through it.
     //
     // A [Charts.LinePointHighlighter] behavior is added manually to enable the
     // highlighting effect.
@@ -47,11 +83,15 @@ class SelectionLineHighlight extends StatelessWidget {
       // set by providing a [dashPattern] or it can be turned off by passing in
       // an empty list. An empty list is necessary because passing in a null
       // value will be treated the same as not passing in a value at all.
+      //
+      // The symbol renderer is configured to render a hollow shape, for
+      // demonstration.
       new charts.LinePointHighlighter(
           showHorizontalFollowLine:
               charts.LinePointHighlighterFollowLineType.none,
           showVerticalFollowLine:
-              charts.LinePointHighlighterFollowLineType.nearest),
+              charts.LinePointHighlighterFollowLineType.nearest,
+          symbolRenderer: new charts.RectSymbolRenderer(isSolid: false)),
       // Optional - By default, select nearest is configured to trigger
       // with tap so that a user can have pan/zoom behavior and line point
       // highlighter. Changing the trigger to tap and drag allows the
@@ -88,4 +128,3 @@ class LinearSales {
 
   LinearSales(this.year, this.sales);
 }
-```