+++ charts_flutter/lib/src/time_series_chart.dart
@@ -13,10 +13,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'dart:collection' show LinkedHashMap;
+
 import 'package:charts_common/common.dart' as common
     show
         AxisSpec,
         DateTimeFactory,
+        NumericAxisSpec,
         Series,
         SeriesRendererConfig,
         TimeSeriesChart;
@@ -42,6 +45,7 @@ class TimeSeriesChart extends CartesianChart<DateTime> {
     common.AxisSpec domainAxis,
     common.AxisSpec primaryMeasureAxis,
     common.AxisSpec secondaryMeasureAxis,
+    LinkedHashMap<String, common.NumericAxisSpec> disjointMeasureAxes,
     common.SeriesRendererConfig<DateTime> defaultRenderer,
     List<common.SeriesRendererConfig<DateTime>> customSeriesRenderers,
     List<ChartBehavior> behaviors,
@@ -58,6 +62,7 @@ class TimeSeriesChart extends CartesianChart<DateTime> {
           domainAxis: domainAxis,
           primaryMeasureAxis: primaryMeasureAxis,
           secondaryMeasureAxis: secondaryMeasureAxis,
+          disjointMeasureAxes: disjointMeasureAxes,
           defaultRenderer: defaultRenderer,
           customSeriesRenderers: customSeriesRenderers,
           behaviors: behaviors,
@@ -76,7 +81,8 @@ class TimeSeriesChart extends CartesianChart<DateTime> {
     return new common.TimeSeriesChart(
         layoutConfig: layoutConfig?.commonLayoutConfig,
         primaryMeasureAxis: primaryMeasureAxis?.createAxis(),
-        secondaryMeasureAxis: secondaryMeasureAxis?.createAxis());
+        secondaryMeasureAxis: secondaryMeasureAxis?.createAxis(),
+        disjointMeasureAxes: createDisjointMeasureAxes());
   }
 
   @override